% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{aapl}
\alias{aapl}
\title{Apple stock data}
\usage{
aapl
}
\description{
Weekly stock data for AAPL (Apple, Inc.) from 1984 to 2012.
}
\section{Variables}{

\itemize{
  \item date
  \item adj_price: Price, adjusted for splits and dividends.
}
}

