\docType{data}
\name{anthoming}
\alias{anthoming}
\title{Homing in desert ants}
\source{
  Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. &
  Ostrowski, E. (1994), \emph{A Handbook of Small Data
  Sets}, Chapman & Hall.

  Duelli, P. and Wehner, R. (1973), The spectal sensitivity
  of polarized light orientation in \emph{Cataglyphis
  bicolor} (Formicidae, Hymenoptera).  \emph{Journal of
  Comparative Physiology}, \bold{86}, 36-53.
}
\description{
  Data from an experiment on the homing performance of a
  desert ant, \emph{Cataglyphis bicolor}.
}
\section{Variables}{
  \itemize{ \item angle: Angle between true home direction
  and the direction that the ant went in (positive is
  clockwise).  \item expt: Number of ants in the
  experimental condition that went in this direction.
  \item ctrl: Number of ants in the control condition that
  went in this direction. }
}
\keyword{datasets}

