% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{corneas}
\alias{corneas}
\title{Corneal thickness of eyes}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994),
  \emph{A Handbook of Small Data Sets}, Chapman & Hall.

  Ehlers, N. On corneal thickness and introcular pressure, II. (1970).
  \emph{Acta Opthalmologica}, \bold{48}, 1107-1112.
}
\usage{
corneas
}
\description{
Corneal thickness of eight people who had glaucoma in one eye.
}
\section{Variables}{

\itemize{
  \item affected Corneal thickness (in microns) of eye affected by glaucoma.
  \item notaffected Corneal thickness (in microns) of eye not affected by glaucoma.
}
}

