% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{drunk}
\alias{drunk}
\title{Convictions for drunkenness}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994),
  \emph{A Handbook of Small Data Sets}, Chapman & Hall.

  Cook, T. (1971). \emph{New Society}, 20 May, 1971.
}
\usage{
drunk
}
\description{
Number of people convicted for drunkenness at Tower Bridge and Lambeth
Magistrates' Courts from January 1 to June 27, 1970, classified by age
and sex.
}
