% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{package}
\name{gcookbook}
\alias{gcookbook}
\alias{package-gcookbook}
\alias{gcookbook-package}
\title{gcookbook: Data sets for "R Graphics Cookbook"}
\description{
This package contains data sets used in the book "R Graphics Cookbook"
by Winston Chang, published by O'Reilly Media.
}
