% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{heightweight}
\alias{heightweight}
\title{Height and weight of schoolchildren}
\source{
Lewis, T., & Taylor, L.R. (1967),
  \emph{Introduction to Experimental Ecology}, Academic Press.
}
\usage{
heightweight
}
\description{
Height and weight of schoolchildren
}
\section{Variables}{

\itemize{
  \item sex
  \item ageYear: Age in years.
  \item ageMonth: Age in months.
  \item heightIn: Height in inches.
  \item weightLb: Weight in pounds.
}
}

