\docType{data}
\name{madmen}
\alias{madmen}
\title{Successful sexual relations in Mad Men (TV show)}
\source{
  Wired Magazine 20.02, February 2012
}
\description{
  Each row of this data frame represents a pair of
  characters who had a sexual relationship on the TV show
  Mad Men, as of the end of season 4. This data can be
  displayed with an undirected graph.
}
\section{Variables}{
  The placement of names in column Name1 as opposed to
  Name2 is arbitrary, and not meaningful. In other words,
  for any row, you can swap the values of Name1 and Name2,
  and it will represent the same information.

  \itemize{ \item Name1: Name of one sexual partner.  \item
  Name2: Name of another sexual partner. }
}
\seealso{
  For a list of attempted sexual pairings, see
  \code{\link{madmen2}}.
}
\keyword{datasets}

