\docType{data}
\name{madmen2}
\alias{madmen2}
\title{Attempted sexual relations in Mad Men (TV show)}
\source{
  Wired Magazine 20.02, February 2012
}
\description{
  Each row of this data frame represents a pair of
  characters on the TV show Mad Men, as of the end of
  season 4. Each row represents an attempted sexual
  relation: the character in the first column, Name1,
  attempted to have sex with the character in the second
  column, Name2. If the relationship goes in both
  directions (the characters had sex with each other), then
  there will be two rows, representing each direction. This
  data can be displayed with a directed graph.
}
\section{Variables}{
  \itemize{ \item Name1: Character who made sexual
  advances.  \item Name2: Character who was the target of
  sexual advances. }
}
\seealso{
  For a list of successful sexual pairings, see
  \code{\link{madmen}}.
}
\keyword{datasets}

