% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{marathon}
\alias{marathon}
\title{Marathon and half-marathon times}
\source{
Downey, A.B. (2011), \emph{Think Stats}, O'Reilly Media.
}
\usage{
marathon
}
\description{
This data set contains mrathon and half-marathon running times for 520
people. Each row represents one person's times.
}
\section{Variables}{

\itemize{
  \item Half: Time in minutes, for half marathon.
  \item Full: Time in minutes, for full marathon.
}
}

