\docType{data}
\name{pg_mean}
\alias{pg_mean}
\title{Means of results from an experiment on plant growth}
\source{
  Dobson, A. J. (1983) \emph{An Introduction to Statistical
  Modelling}, Chapman & Hall.
}
\description{
  This data set simply has groupwise means of the
  \code{\link{PlantGrowth}} data set. The purpose of this
  summarized data set is to make it easy to use for example
  graphs.
}
\seealso{
  The source data set, \code{\link{PlantGrowth}}.
}
\keyword{datasets}

