% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{plum_wide}
\alias{plum_wide}
\title{Plum root cuttings (wide format)}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994),
  \emph{A Handbook of Small Data Sets}, Chapman & Hall.

 Bartlett, M.S. (1935), Contingency table interactions,
 \emph{Journal of the Royal Statistical Society Supplement}, \bold{2}, 248-252.
}
\usage{
plum_wide
}
\description{
This is data from an experiment to investigate the effect of cutting length
and planting time on the survival of plum root cuttings.
}
\section{Variables}{

\itemize{
  \item length: Cutting length.
  \item time: Planting time.
  \item dead: Number of dead plants in this condition.
  \item alive: Number of alive plants in this condition.
}
}

\seealso{
This data frame is in "wide" format. See \code{\link{plum}} for
  the same data in "long" format.
}
