% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{simpledat}
\alias{simpledat}
\title{Simple example data set}
\usage{
simpledat
}
\description{
This data set is for examples of R graphics.
}
\seealso{
This data frame is in "wide" format. See \code{\link{simpledat_long}}
  for the same data in "long" format.
}
