% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{tg}
\alias{tg}
\title{Summarized ToothGrowth data}
\usage{
tg
}
\description{
Summarized ToothGrowth data
}
\section{Variables}{

\itemize{
  \item supp
  \item dose
  \item length
}
}

\seealso{
\code{\link[datasets]{ToothGrowth}}
}
