\docType{data}
\name{uspopchange}
\alias{uspopchange}
\title{Change in population of states in the U.S. between 2000 and 2010}
\source{
  U.S. Census Bureau, Statistical Abstract of the United
  States, 2012, Table 14.
  \url{http://www.census.gov/compendia/statab/2012/tables/12s0014.pdf}
}
\description{
  This data set represents the percent change in population
  of states in the U.S. from 2000 to 2010.
}
\section{Variables}{
  \itemize{ \item State \item Abb: Abbreviated state name.
  \item Region: Region of country that the state is in.
  \item Change: Percent change in population. }
}
\keyword{datasets}

