% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{wind}
\alias{wind}
\title{Wind speed and direction}
\source{
Great Lakes Environmental Research Laboratory:
    \url{http://www.glerl.noaa.gov/metdata/chi/}
}
\usage{
wind
}
\description{
This data set contains the wind speed and direction over the course of a
single day in Chicago (February 29, 2012).
}
\section{Variables}{

\itemize{
  \item TimeUTC: Time of day in minutes; 0 is midnight.
  \item Temp: Temperature in Celcius.
  \item WindAvg: Average wind speed in m/s in this time block.
  \item WindMax: Maximum wind speed in m/s in this time block.
  \item WindDir: Average direction that wind comes from (0=north, 90=east).
  \item SpeedCat: Average wind speed, categorized in 5 m/s groups.
  \item DirCat: Average direction, categorized in 15-degree groups.
}
}

