% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{cabbage_exp}
\alias{cabbage_exp}
\title{Summary of cabbages data set}
\usage{
cabbage_exp
}
\description{
This data set has groupwise means, standard deviations, counts, and standard
error of the mean for the \code{\link{cabbages}} data set from the MASS
package. The purpose of this summarized data set is to make it easy to use
for example graphs.
}
\seealso{
The source data set in the MASS package, \code{\link{cabbages}}.
}
