% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{climate}
\alias{climate}
\title{Global climate temperature anomaly data from 1800 to 2011}
\source{
Berkeley Earth Project: \url{http://berkeleyearth.org/land-and-ocean-data/}

  Climatic Research Unit (CRUTEM3): \url{https://crudata.uea.ac.uk/cru/data/crutem3/}

  NASA: \url{https://data.giss.nasa.gov/gistemp/}
}
\usage{
climate
}
\description{
This data set includes estimated global temperature anamoly data for the
years 1800 through 2011. The anomaly is the difference from the baseline
temperature, which is the mean of the yearly temperatures from 1951-1980.
}
\section{Variables}{

\itemize{
  \item Source: Data source (Berkeley, CRUTEM3, NASA).
  \item Year: Year for the estimate.
  \item Anomaly1y: Temperature anomaly in Celcius, smoothed over one year.
  \item Anomaly5y: Temperature anomaly in Celcius, smoothed over five years.
  \item Anomaly10y: Temperature anomaly in Celcius, smoothed over ten years.
  \item Unc10y: Uncertainty for 10-year-smoothed anomaly.
}
}

