% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{countries}
\alias{countries}
\title{Health and economic data about countries around the world from 1960-2010}
\source{
World Bank: \url{https://data.worldbank.org/}
}
\usage{
countries
}
\description{
Health and economic data about countries around the world from 1960-2010,
from the World Bank.
}
\section{Variables}{

\itemize{
  \item Name: Name of country
  \item Code: Short country code
  \item Year
  \item GDP: Per capita Gross Domestic Product, in adjusted 2011 U.S. Dollars
  \item laborrate: Labor rate.
  \item healthexp: Health expenditures in U.S. Dollars.
  \item infmortality: Infant mortality per 1000 live births.
}
}

