\docType{package}
\name{gcookbook}
\alias{gcookbook}
\alias{gcookbook-package}
\alias{package-gcookbook}
\title{gcookbook: Data sets for "R Graphics Cookbook"}
\description{
  This package contains data sets used in the book "R
  Graphics Cookbook" by Winston Chang, published by
  O'Reilly Media.
}

