% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{plum}
\alias{plum}
\title{Plum root cuttings (long format)}
\source{
Hand, D.J., Daly, F., Lunn, A.D., McConway, K.J. & Ostrowski, E. (1994),
  \emph{A Handbook of Small Data Sets}, Chapman & Hall.

 Bartlett, M.S. (1935), Contingency table interactions,
 \emph{Journal of the Royal Statistical Society Supplement}, \bold{2}, 248-252.
}
\usage{
plum
}
\description{
This is data from an experiment to investigate the effect of cutting length
and planting time on the survival of plum root cuttings.
}
\section{Variables}{

\itemize{
  \item length: Cutting length.
  \item time: Planting time.
  \item survival: Survival status.
  \item count: Number of plants.
}
}

\seealso{
This data frame is in "long" format. See \code{\link{plum_wide}} for
  the same data in "wide" format.
}
