% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{simpledat_long}
\alias{simpledat_long}
\title{Simple example data set (long format)}
\usage{
simpledat_long
}
\description{
This data set is for examples of R graphics.
}
\seealso{
This data frame is in "long" format. See \code{\link{simpledat}}
  for the same data in "wide" format.
}
