% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{uspopage}
\alias{uspopage}
\title{Age distribution of population in the United States, 1900-2002}
\source{
U.S. Census Bureau, Statistical Abstract of the United States, 2003, HS-3:
    \url{http://www.census.gov/statab/hist/HS-03.pdf} and
    \url{www.census.gov/statab/hist/02HS0003.xls}.
}
\usage{
uspopage
}
\description{
These are the estimated (not counted) values by the U.S. Census.
}
\section{Variables}{

\itemize{
  \item Year
  \item AgeGroup
  \item Thousands: Number of people, in thousands.
}
}

