% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{uspopage}
\alias{uspopage}
\title{Age distribution of population in the United States, 1900-2002}
\source{
U.S. Census Bureau, Statistical Abstract of the United States, 2003, HS-3:
    \url{http://www2.census.gov/library/publications/2004/compendia/statab/123ed/hist/hs-03.pdf}
}
\usage{
uspopage
}
\description{
These are the estimated (not counted) values by the U.S. Census.
}
\section{Variables}{

\itemize{
  \item Year
  \item AgeGroup
  \item Thousands: Number of people, in thousands.
}
}

