% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcookbook.r
\docType{data}
\name{worldpop}
\alias{worldpop}
\title{World population estimates from 10,000 B.C. to 2,000 A.D.}
\usage{
worldpop
}
\description{
World population estimates from 10,000 B.C. to 2,000 A.D.
}
\section{Variables}{

\itemize{
  \item Year
  \item Population: Estimated population, in thousands
}
}

